BUILD127: fix V3 residual resize integer typing

+++ lib/hcv_ml_v3_photo_residual.dart
@@ -169,8 +169,10 @@ class HCVMLV3PhotoResidual {
         (oriented.width > oriented.height
             ? oriented.width
             : oriented.height);
-    final width = (oriented.width * scale).round().clamp(1, imageSize);
-    final height = (oriented.height * scale).round().clamp(1, imageSize);
+    final width =
+        (oriented.width * scale).round().clamp(1, imageSize).toInt();
+    final height =
+        (oriented.height * scale).round().clamp(1, imageSize).toInt();
     final resized = img.copyResize(
       oriented,
       width: width,
